# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * User Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthResponse {
  token: string;
  user: User;
}