# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * User Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  /** @minLength 1 */
  name?: string;
  bio?: string;
  avatarUrl?: string;
}